pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "daily-python-practice"
version = "0.1.0"
description = "Repository for daily Python practice exercises and learning activities"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Antonio Alcaide Moreno", email = "antonion1995@gmail.com"}
]
keywords = ["python", "exercises", "practice", "learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.hatch.build]
# Point at your src/ folder as the root of your packages:
source-directories = ["src"]
dev-mode-dirs = ["src"]

[tool.hatch.build.targets.wheel]
# Refer to the importable package name, not the on-disk path:
packages = ["src/exercises"]

[tool.hatch.build.targets.editable]
# Same source root for editable installs:
path = "src"

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "jupyter>=1.0.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/AntonioAlcaideMoreno/daily-python-practice"
Repository = "https://github.com/AntonioAlcaideMoreno/daily-python-practice.git"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--cov=src",
    "--import-mode=importlib",
    "--strict-markers",
    "-v",
]
testpaths = [
    "tests",           # Root-level tests directory
    "src/tests",       # Tests inside src directory
]
pythonpath = ["src"]

[project.scripts]
exercise-0001 = "exercises.recursive_nested_list_function:main"
exercise-0002 = "exercises.fibonaccy:main"
